+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - PolygonScan parser: failed transactions should not result in Deposits.
 - SnowTrace parser: failed transactions should not result in Deposits.
 - Etherscan parser: treat NFTs as unique assets. ([#234](https://github.com/BittyTax/BittyTax/issues/234))
+- Trezor Suite parser: support "SELF" and "FAILED" transaction types.
 ### Added
 - Conversion tool: generic parser added to support import from CryptoCredible tool.
 - Conversion tool: added parser for SnowTrace explorer.